requirements.txt
keras
streamlit 
pillow 
numpy
tensorflow

